requirements.txt
# Core
pydantic>=2.0,<3.0
python-dotenv>=1.0,<2.0
httpx>=0.27,<1.0

# NYC Open Data (PLUTO real estate)
sodapy>=2.2,<3.0

# LLM outreach generation — uses Gemini REST API via httpx (no SDK needed)

# PDF text extraction
PyPDF2>=3.0,<4.0

# Dashboard
streamlit>=1.32,<2.0

# Testing
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<1.0
